增加切换wifi和以太网的脚本

+++ update.sh
@@ -1,15 +1,14 @@
-############################################################
-#  git仓库更新到最新的develop分支的代码 
-############################################################
 function update()
 {
-	cd ${1}
-	echo $1
-	git fetch
-	git reset
-	git checkout develop
-	git reset --hard origin/develop
-	cd ..
+	if [ -d "${1}" ] ; then
+		cd ${1}
+		echo $1
+		git fetch
+		git clean -fd
+		git checkout develop
+		git reset --hard origin/develop
+		cd ..
+	fi
 }
 ls | while read LINE; do update $LINE; done
